chore: remove ioc
