+++ apps/wallet/browser/kernel/utils/config.ts
@@ -1,24 +1,22 @@
-export const makeConfig = () => {
-	return {
-		FIREBASE_API_KEY,
-		BUILD_TARGET,
-		BROWSER_CLIENT_ID,
-		EXTENSION_CLIENT_ID,
-		FIREBASE_AUTH_DOMAIN,
-		FIREBASE_PROJECT_ID,
-		FIREBASE_STORAGE_BUCKET,
-		FIREBASE_MESSAGING_SENDER_ID,
-		FIREBASE_APP_ID,
-		FIREBASE_MEASUREMENT_ID,
-		FIREBASE_VAPID_KEY,
-		GATEFI_ENDPOINT,
-		GATEFI_MERCHANT_ID,
-		WEB3AUTH_ID,
-		GRAPHQL_ENDPOINT,
-		PIXEVERSE_ENDPOINT,
-		PIXEVERSE_ORIGIN,
-		PIXEVERSE_URL,
-		SOLANA_CLUSTER_URL,
-		GASILON_ENDPOINT,
-	};
+export const environment = {
+	FIREBASE_API_KEY,
+	BUILD_TARGET,
+	BROWSER_CLIENT_ID,
+	EXTENSION_CLIENT_ID,
+	FIREBASE_AUTH_DOMAIN,
+	FIREBASE_PROJECT_ID,
+	FIREBASE_STORAGE_BUCKET,
+	FIREBASE_MESSAGING_SENDER_ID,
+	FIREBASE_APP_ID,
+	FIREBASE_MEASUREMENT_ID,
+	FIREBASE_VAPID_KEY,
+	GATEFI_ENDPOINT,
+	GATEFI_MERCHANT_ID,
+	WEB3AUTH_ID,
+	GRAPHQL_ENDPOINT,
+	PIXEVERSE_ENDPOINT,
+	PIXEVERSE_ORIGIN,
+	PIXEVERSE_URL,
+	SOLANA_CLUSTER_URL,
+	GASILON_ENDPOINT,
 };
