+++ apps/wallet/browser/kernel/utils/init.ts
@@ -1,0 +1,7 @@
+import { configure } from '@walless/store';
+
+import { storage } from './storage';
+
+export const initModules = async () => {
+	await Promise.all([configure(storage)]);
+};
