+++ apps/wallet/browser/kernel/utils/storage.ts
@@ -1,0 +1,4 @@
+import { create } from '@walless/store';
+import IDBPouch from 'pouchdb-adapter-idb';
+
+export const storage = create('engine', IDBPouch);
